cetus-distributor/pyproject.toml
[tool.poetry]
name = "cetus-distributor"
version = "0.1.0"
description = "Cetus Distributor"
authors = ["Vitaliy Boyarsky <boyarsky.vitaliy@live.com>"]
packages = [
    { include = "cetus_distributor", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
pyxis = { git = "https://github.com/vitalibo/pyxis.git", tag = "0.2.1-pyspark" }
pydantic = "^2.5.3"
pyyaml = "^6.0.1"
aws-cdk-lib = "^2.147.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pylint = "^3.0.3"
isort = "^5.13.2"
poethepoet = "^0.24.4"
pyspark = "^3.3.0"
toml = "^0.10.2"

[tool.poe]
poetry_command = ""

[tool.poe.tasks]
isort = { cmd = "isort ./src/ ./tests/", help = "Sort imports alphabetically." }
style = { cmd = "pylint ./src/ ./tests/ --rcfile=../.pylintrc", help = "Run static code analyser.", deps = ["isort"] }
test = { cmd = "pytest -v -p no:cacheprovider --disable-warnings ./tests/", help = "Run unit tests.", env = { PYTHONPATH = "./src" } }
clean = { cmd = "rm -rf ./.pytest_cache ./build ./dist ./src/*.egg-info", help = "Clean workdir." }

[tool.isort]
line_length = 120
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
